refactor: extract the layout into its own module
First half of feature 2. No behaviour change: the recursion, the sub-rect
calculation and the headerBoxes filtering are the same code, moved out of
the controller behind a typed signature.

The controller no longer contains any layout logic, which is what makes
the pipeline change reviewable as a diff of its own.
